requirements.txt
aniso8601==7.0.0
argcomplete==1.9.3
boto3==1.9.227
botocore==1.12.227
certifi==2019.9.11
cfn-flip==1.2.1
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.2.4
django-cors-headers==3.1.0
django-filter==2.2.0
django-rest-auth==0.9.5
djangorestframework==3.10.2
docutils==0.15.2
durationpy==0.5
future==0.16.0
graphene==2.1.8
graphene-django==2.5.0
graphql-core==2.2.1
graphql-relay==2.0.0
gunicorn==19.9.0
hjson==3.0.1
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
Markdown==3.1.1
MarkupSafe==1.1.1
placebo==0.9.0
promise==2.2.1
psycopg2==2.8.3
Pygments==2.4.2
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
Rx==1.6.1
s3transfer==0.2.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
toml==0.10.0
tqdm==4.19.1
troposphere==2.5.1
Unidecode==1.1.1
uritemplate==3.0.0
urllib3==1.25.3
Werkzeug==0.15.6
whitenoise==4.1.3
wsgi-request-logger==0.4.6
zappa==0.48.2
